Use tile+fuse on CPU side.
Use tile+fuse on CPU side as well by modifying the CPU implementation
to match the GPU implementation. Main changes needed were to change
the LaunchConfig to be made independent of the target architecture.
