+++ iree/compiler/Conversion/Common/BUILD
@@ -22,20 +22,29 @@ cc_library(
     name = "Common",
     srcs = [
         "DeclareNumWorkgroupsFnPass.cpp",
+        "LaunchConfig.cpp",
         "LegalizeNumWorkgroupsFnPass.cpp",
+        "Transforms.cpp",
     ],
     hdrs = [
         "Attributes.h",
+        "LaunchConfig.h",
         "Passes.h",
+        "Transforms.h",
     ],
     deps = [
         "//iree/compiler/Conversion/CodegenUtils",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/IREE/IR",
         "//iree/compiler/Dialect/Shape/IR",
+        "@llvm-project//llvm:Support",
         "@llvm-project//mlir:CFGTransforms",
+        "@llvm-project//mlir:GPUDialect",
         "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:LinalgOps",
+        "@llvm-project//mlir:LinalgTransforms",
         "@llvm-project//mlir:Pass",
+        "@llvm-project//mlir:SPIRVLowering",
         "@llvm-project//mlir:StandardOps",
         "@llvm-project//mlir:Transforms",
         "@org_tensorflow//tensorflow/compiler/mlir/hlo",
