+++ iree/compiler/Conversion/Common/Transforms.cpp
@@ -1,0 +1,260 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//===- Transforms.cpp - Transformations common to all backends ------------===//
+//
+// Implements transformations that are common to all backends.
+//
+//===----------------------------------------------------------------------===//
+
+#include "iree/compiler/Conversion/Common/Transforms.h"
+
+#include "iree/compiler/Conversion/CodegenUtils/FunctionUtils.h"
+#include "iree/compiler/Conversion/CodegenUtils/GetNumWorkgroups.h"
+#include "iree/compiler/Conversion/CodegenUtils/MarkerUtils.h"
+#include "iree/compiler/Conversion/CodegenUtils/MatmulCodegenStrategy.h"
+#include "iree/compiler/Conversion/Common/Attributes.h"
+#include "mlir/Dialect/GPU/GPUDialect.h"
+#include "mlir/Dialect/Linalg/Analysis/DependenceAnalysis.h"
+#include "mlir/Dialect/Linalg/IR/LinalgOps.h"
+#include "mlir/Dialect/Linalg/Utils/Utils.h"
+#include "mlir/IR/Function.h"
+#include "mlir/IR/PatternMatch.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+
+#define DEBUG_TYPE "iree-linalg-tile-and-fuse"
+
+namespace mlir {
+namespace iree_compiler {
+
+//===----------------------------------------------------------------------===//
+// Utility functions
+//===----------------------------------------------------------------------===//
+
+/// Apply canonicalizations related to tiling to make promotion/vectorization
+/// easier.
+void applyCanonicalizationPatterns(MLIRContext *context, Operation *op) {
+  OwningRewritePatternList canonicalizationPatterns;
+  canonicalizationPatterns.insert<AffineMinCanonicalizationPattern>(context);
+  AffineApplyOp::getCanonicalizationPatterns(canonicalizationPatterns, context);
+  AffineMinOp::getCanonicalizationPatterns(canonicalizationPatterns, context);
+  SubViewOp::getCanonicalizationPatterns(canonicalizationPatterns, context);
+  applyPatternsAndFoldGreedily(op, std::move(canonicalizationPatterns));
+}
+
+/// Once fused the fused views that are due to a RAW dependence can be promoted
+/// to workgroup memory. This will make the intermediate storage dead.
+static LogicalResult promoteFusedViews(OpBuilder &builder,
+                                       ArrayRef<linalg::LinalgOp> fusedOps,
+                                       const TileAndFuseOptions &options) {
+  linalg::Aliases aliases;
+  linalg::LinalgDependenceGraph dependenceGraph(aliases, fusedOps);
+  auto fusableDependences =
+      linalg::findAllFusableDependences(fusedOps, dependenceGraph);
+
+  DenseSet<Value> promotedViews;
+  OpBuilder::InsertionGuard guard(builder);
+  builder.setInsertionPoint(*fusedOps.begin());
+  for (linalg::LinalgOp op : llvm::reverse(fusedOps)) {
+    auto dependences = fusableDependences.lookup(op);
+    if (dependences.empty()) continue;
+    if (!llvm::hasSingleElement(dependences)) {
+      return op.emitError(
+          "unable to promote ops with multiple fusable dependences");
+    }
+    auto dependence = dependences.front();
+    unsigned producerIdx = dependence.dependentOpView.operandIndex;
+    linalg::LinalgOp consumer =
+        cast<linalg::LinalgOp>(dependence.indexingOpView.op);
+    unsigned consumerIdx = dependence.indexingOpView.operandIndex;
+    Value consumerView = consumer.getShapedOperand(consumerIdx);
+    Value promotedView = nullptr;
+
+    if (promotedViews.count(consumerView)) {
+      promotedView = consumerView;
+    } else if (dependence.dependenceType ==
+               linalg::LinalgDependenceGraph::RAW) {
+      Optional<linalg::PromotionInfo> promotionInfo =
+          linalg::promoteSubviewAsNewBuffer(
+              builder, op.getLoc(),
+              op.getShapedOperand(producerIdx).getDefiningOp<SubViewOp>(),
+              options.allocationFn);
+      if (!promotionInfo) {
+        return op.emitError("unable to promote RAW dependence");
+      }
+      promotedView = promotionInfo->partialLocalView;
+      consumer.getOperation()->setOperand(consumerIdx, promotedView);
+      promotedViews.insert(promotedView);
+    }
+    if (!promotedView) continue;
+    op.getOperation()->setOperand(producerIdx, promotedView);
+  }
+  return success();
+}
+
+/// Tile+Fuse only tiles the loops that can be fused. Tile any of the unfused
+/// loops in the operation based on the configuration.
+static linalg::LinalgOp tileUnfusedLoops(
+    OpBuilder &builder, linalg::LinalgOp linalgOp,
+    const std::set<unsigned> &fusedLoopDims, const LaunchConfig &launchConfig) {
+  SmallVector<int64_t, 4> tileSizes =
+      llvm::to_vector<4>(launchConfig.getTileSizes(linalgOp, 0));
+  tileSizes.resize(linalgOp.getNumLoops(), 0);
+  for (unsigned loopNum : fusedLoopDims) {
+    tileSizes[loopNum] = 0;
+  }
+  if (llvm::all_of(tileSizes, [](int64_t v) { return !v; })) return linalgOp;
+  Optional<linalg::TiledLinalgOp> tiledOp = tileLinalgOp(
+      builder, linalgOp,
+      linalg::LinalgTilingOptions().setTileSizes(tileSizes).setLoopType(
+          linalg::LinalgTilingLoopType::ParallelLoops));
+  if (!tiledOp) return nullptr;
+  linalgOp.erase();
+  return tiledOp->op;
+}
+
+/// Main utility function that implements the tile and fuse.
+static Optional<linalg::TiledAndFusedLinalgOps> tileAndFuseLinalgOps(
+    OpBuilder &builder, FuncOp funcOp, ArrayRef<linalg::LinalgOp> fusableOps,
+    const linalg::LinalgDependenceGraph &dependenceGraph,
+    const LaunchConfig &launchConfig, const TileAndFuseOptions &options) {
+  // Get the tile sizes to use from the last fusable op and the tile+fuse all
+  // ops.
+  SmallVector<int64_t, 4> tileSizes =
+      llvm::to_vector<4>(launchConfig.getTileSizes(fusableOps.back(), 0));
+  linalg::LinalgTilingOptions tilingOptions;
+  tilingOptions.setDistributionOptions(options.distributionOptions)
+      .setTileSizes(tileSizes)
+      .setLoopType(linalg::LinalgTilingLoopType::ParallelLoops);
+
+  Optional<linalg::TiledAndFusedLinalgOps> tiledAndFusedOps = llvm::None;
+  if (fusableOps.size() == 1) {
+    linalg::LinalgOp linalgOp = fusableOps.front();
+    Optional<linalg::TiledLinalgOp> tiledOp =
+        tileLinalgOp(builder, linalgOp, tilingOptions);
+    if (!tiledOp) {
+      linalgOp.emitError("unable to tile operation");
+      return llvm::None;
+    }
+    tiledAndFusedOps = linalg::TiledAndFusedLinalgOps{tiledOp->op, {}, {}, {}};
+    auto seq = llvm::seq<unsigned>(0, tileSizes.size());
+    tiledAndFusedOps->fusedLoopDims.insert(seq.begin(), seq.end());
+    tiledAndFusedOps->fusedLoops.assign(tiledOp->loops.begin(),
+                                        tiledOp->loops.end());
+  } else {
+    tiledAndFusedOps = tileAndFuseLinalgOps(builder, fusableOps,
+                                            dependenceGraph, tilingOptions);
+  }
+  if (!tiledAndFusedOps) {
+    funcOp.emitError("tile and fuse of linalg operations failed");
+    return llvm::None;
+  }
+
+  // Update the launch configuration.
+  SmallVector<unsigned, 2> distributedLoops =
+      llvm::to_vector<2>(tiledAndFusedOps->fusedLoopDims);
+  if (funcOp.getAttr(getNumWorkgroupsFnAttrName()) &&
+      failed(createNumWorkgroupsFromResultShape(
+          builder, fusableOps.back(), funcOp, getNumWorkgroupsFnAttrName(),
+          tileSizes, distributedLoops))) {
+    funcOp.emitError("failed to update launch configuration");
+    return llvm::None;
+  }
+
+  // Delete all the original operations.
+  for (auto linalgOp : fusableOps) linalgOp.erase();
+
+  // Add workgroup markers to all the tiled and fused operations.
+  for (auto fusedProducer : tiledAndFusedOps->fusedProducers) {
+    setMarker(fusedProducer, getWorkgroupMarker());
+  }
+  setMarker(tiledAndFusedOps->op, getWorkgroupMarker());
+
+  return tiledAndFusedOps;
+}
+
+LogicalResult tileAndFuseLinalgBufferOps(
+    FuncOp funcOp, ArrayRef<linalg::LinalgOp> linalgOps,
+    const linalg::LinalgDependenceGraph &dependenceGraph,
+    const LaunchConfig &launchConfig, const TileAndFuseOptions &options) {
+  // Collect all operations that are to be tiled-and-fused.
+  MLIRContext *context = funcOp.getContext();
+  SmallVector<linalg::LinalgOp, 4> fusableOps;
+  for (Operation *operation : linalgOps) {
+    if (!launchConfig.hasTileSizes(operation)) continue;
+    fusableOps.push_back(cast<linalg::LinalgOp>(operation));
+  }
+  if (fusableOps.empty()) return success();
+
+  OpBuilder builder(context);
+  Optional<linalg::TiledAndFusedLinalgOps> tiledAndFusedOps =
+      tileAndFuseLinalgOps(builder, funcOp, fusableOps, dependenceGraph,
+                           launchConfig, options);
+  if (!tiledAndFusedOps) {
+    return funcOp.emitError("failed to tile and fuse operations");
+  }
+
+  LLVM_DEBUG({
+    llvm::dbgs() << "--- After Fusion on buffers ---\n";
+    funcOp.print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
+    llvm::dbgs() << "\n\n";
+  });
+
+  applyCanonicalizationPatterns(context, funcOp);
+
+  LLVM_DEBUG({
+    llvm::dbgs() << "--- After Canonicalization ---\n";
+    funcOp.print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
+    llvm::dbgs() << "\n\n";
+  });
+
+  if (options.allocationFn) {
+    SmallVector<linalg::LinalgOp, 4> promoteFusedViewOps(
+        tiledAndFusedOps->fusedProducers.begin(),
+        tiledAndFusedOps->fusedProducers.end());
+    promoteFusedViewOps.push_back(tiledAndFusedOps->op);
+
+    if (failed(promoteFusedViews(builder, promoteFusedViewOps, options))) {
+      return failure();
+    }
+  }
+
+  LLVM_DEBUG({
+    llvm::dbgs() << "--- After Promotion ---\n";
+    funcOp.print(llvm::dbgs(), OpPrintingFlags().useLocalScope());
+    llvm::dbgs() << "\n\n";
+  });
+
+  // Tile the unfused loops. Set the tile sizes for the fused loops to be zero
+  // to avoid tiling them again.
+  for (linalg::LinalgOp &fusedOp : tiledAndFusedOps->fusedProducers) {
+    linalg::LinalgOp tiledOp = tileUnfusedLoops(
+        builder, fusedOp, tiledAndFusedOps->fusedLoopDims, launchConfig);
+    if (!tiledOp) {
+      return fusedOp.emitError("unable to tile unfused loops");
+    }
+  }
+  linalg::LinalgOp tiledOp =
+      tileUnfusedLoops(builder, tiledAndFusedOps->op,
+                       tiledAndFusedOps->fusedLoopDims, launchConfig);
+  if (!tiledOp) {
+    return tiledAndFusedOps->op.emitError("unable to tile unfused loops");
+  }
+
+  applyCanonicalizationPatterns(context, funcOp);
+  return success();
+}
+
+}  // namespace iree_compiler
+}  // namespace mlir
