+++ iree/compiler/Conversion/LinalgToLLVM/LinalgTileAndDistributePass.cpp
@@ -17,6 +17,7 @@
 #include "iree/compiler/Conversion/CodegenUtils/MarkerUtils.h"
 #include "iree/compiler/Conversion/CodegenUtils/MatmulCodegenStrategy.h"
 #include "iree/compiler/Conversion/Common/Attributes.h"
+#include "iree/compiler/Conversion/Common/Transforms.h"
 #include "iree/compiler/Conversion/LinalgToLLVM/KernelDispatch.h"
 #include "iree/compiler/Dialect/IREE/IR/IREEDialect.h"
 #include "iree/compiler/Dialect/IREE/IR/IREEOps.h"
@@ -41,11 +42,6 @@ struct LinalgTileAndDistributePass
   LinalgTileAndDistributePass() = default;
   LinalgTileAndDistributePass(const LinalgTileAndDistributePass &pass) {}
   void runOnOperation() override;
-
- private:
-  ListOption<int64_t> tileSizes{
-      *this, "tile-sizes", llvm::cl::desc("Set tile sizes to use"),
-      llvm::cl::ZeroOrMore, llvm::cl::MiscFlags::CommaSeparated};
 };
 }  // namespace
 
@@ -147,57 +143,57 @@ void LinalgTileAndDistributePass::runOnOperation() {
        linalg::DistributionMethod::CyclicNumProcsEqNumIters,
        linalg::DistributionMethod::CyclicNumProcsEqNumIters}};
 
-  CPUKernelDispatch cpuKernelDispatch;
-
   for (FuncOp funcOp : module.getOps<FuncOp>()) {
     if (!isEntryPoint(funcOp)) continue;
 
-    // Compute the Linalg Dependence Graph.
+    Region &body = funcOp.getBody();
+    if (!llvm::hasSingleElement(body.getBlocks())) {
+      funcOp.emitError("unhandled dispatch function with multiple blocks");
+      return signalPassFailure();
+    }
+    Block &block = body.front();
+    auto linalgOps = block.getOps<linalg::LinalgOp>();
+    if (linalgOps.empty()) continue;
+
+    SmallVector<linalg::LinalgOp, 4> linalgOpsVec =
+        llvm::to_vector<4>(llvm::map_range(linalgOps, [](Operation *op) {
+          return cast<linalg::LinalgOp>(op);
+        }));
     linalg::Aliases aliases;
-    linalg::LinalgDependenceGraph dependenceGraph =
-        linalg::LinalgDependenceGraph::buildDependenceGraph(aliases, funcOp);
+    linalg::LinalgDependenceGraph dependenceGraph(aliases, linalgOpsVec);
+    Optional<LaunchConfig> launchConfigOpt =
+        initCPULaunchConfig(context, dependenceGraph, linalgOpsVec);
+    if (!launchConfigOpt) {
+      funcOp.emitError("unable to find launch configuration");
+      return signalPassFailure();
+    }
+    LaunchConfig &launchConfig = *launchConfigOpt;
 
-    OwningRewritePatternList patterns;
-
-    auto linalgTilingOptions =
-        linalg::LinalgTilingOptions()
-            .setDistributionOptions(workgroupDistributionOptions)
-            .setLoopType(linalg::LinalgTilingLoopType::ParallelLoops);
-    tileSizes.empty()
-        ? linalgTilingOptions.setTileSizeComputationFunction(
-              [&cpuKernelDispatch](
-                  OpBuilder &builder,
-                  Operation *operation) -> SmallVector<Value, 4> {
-                return TileSizeFn::get<TilingLevel::WorkGroupTiles>(
-                    cpuKernelDispatch, builder, operation);
-              })
-        : linalgTilingOptions.setTileSizes(ArrayRef<int64_t>(tileSizes));
-    patterns.insert<TileAndFuseToCPUThreads<linalg::MatmulOp>,
-                    TileAndFuseToCPUThreads<linalg::BatchMatmulOp>,
-                    TileToCPUThreads<linalg::MatmulOp>,
-                    TileToCPUThreads<linalg::BatchMatmulOp>>(
-        context, dependenceGraph, cpuKernelDispatch, linalgTilingOptions,
-        linalg::LinalgMarker(ArrayRef<Identifier>(),
-                             Identifier::get(getWorkgroupMarker(), context)));
-
-    // Tile and distribute to CPU threads.
-    applyPatternsAndFoldGreedily(funcOp, std::move(patterns));
-
-    // Apply canonicalization patterns.
-    OwningRewritePatternList canonicalizationPatterns;
-    canonicalizationPatterns.insert<AffineMinCanonicalizationPattern>(context);
-    AffineApplyOp::getCanonicalizationPatterns(canonicalizationPatterns,
-                                               context);
-    AffineMinOp::getCanonicalizationPatterns(canonicalizationPatterns, context);
-    SubViewOp::getCanonicalizationPatterns(canonicalizationPatterns, context);
-
-    applyPatternsAndFoldGreedily(funcOp, std::move(canonicalizationPatterns));
-
-    // Delete the ops that are marked for deletion.
-    funcOp.walk([](linalg::LinalgOp linalgOp) {
-      if (hasMarker(linalgOp.getOperation(), getDeleteMarker()))
-        linalgOp.getOperation()->erase();
+    LLVM_DEBUG({
+      llvm::dbgs() << "@func " << funcOp.getName() << "\n";
+      for (auto op : linalgOps) {
+        llvm::dbgs() << "\t" << op.getOperation()->getName() << " : ";
+        TileSizesListTypeRef configTileSizes = launchConfig.getTileSizes(op);
+        llvm::dbgs() << "{";
+        std::string sep = "";
+        for (auto &level : enumerate(configTileSizes)) {
+          llvm::dbgs() << sep << level.index() << " : [";
+          sep = ", ";
+          interleaveComma(level.value(), llvm::dbgs());
+          llvm::dbgs() << "]";
+        }
+        llvm::dbgs() << "}\n";
+      }
     });
+
+    TileAndFuseOptions tileAndFuseOptions = {workgroupDistributionOptions,
+                                             nullptr};
+    if (failed(tileAndFuseLinalgBufferOps(funcOp, linalgOpsVec, dependenceGraph,
+                                          launchConfig, tileAndFuseOptions))) {
+      return signalPassFailure();
+    }
+
+    launchConfig.finalize(funcOp);
   }
 }
